Split MetricReader into partial classes
